requirements.txt
biom-format==1.3.1
biopython==1.63
bz2file==0.98
-e git+https://rschwager-hsph@bitbucket.org/rschwager-hsph/ccfa.git@ab5e834d3e6f596a08012b086b4b95f491c853b4#egg=mibc-master
nose==1.3.0
numpy==1.8.0
pyqi==0.3.1
python-dateutil==2.2
-e git+https://github.com/JoseBlanca/seq_crumbs.git@66ace168aaaffb3b9b87afca34454ee24b441bda#egg=seq_crumbs-master
six==1.4.1
bottle
